+++ src/HearThis/Script/ParatextScriptProvider.cs
@@ -1,7 +1,7 @@
 // --------------------------------------------------------------------------------------------
-#region // Copyright (c) 2015, SIL International. All Rights Reserved.
-// <copyright from='2011' to='2015' company='SIL International'>
-//		Copyright (c) 2015, SIL International. All Rights Reserved.
+#region // Copyright (c) 2018, SIL International. All Rights Reserved.
+// <copyright from='2011' to='2018' company='SIL International'>
+//		Copyright (c) 2018, SIL International. All Rights Reserved.
 //
 //		Distributable under the terms of the MIT License (http://sil.mit-license.org/)
 // </copyright>
@@ -27,7 +27,7 @@ namespace HearThis.Script
 		private HashSet<string> _allEncounteredParagraphStyleNames; // This will not include the ones that are always ignored.
 		private IBibleStats _versificationInfo;
 
-		// These are markers that ARE paragraph and IsPublishableVernacular, but we don't want to read them.
+		// These are markers that ARE paragraph and IsPublishable, but we don't want to read them.
 		// They should be followed by a single text node that will be skipped too.
 		private readonly HashSet<string> _furtherParagraphIgnorees = new HashSet<string> { "id", "h", "h1", "h2", "h3", "r", "toc1", "toc2", "toc3", "io1","io2","io3" };
 
@@ -186,9 +186,10 @@ namespace HearThis.Script
 				previousMarker = t?.Marker ?? previousMarker;
 
 				t = tokens[i];
+
 				state.UpdateState(tokens, i);
 
-				if (!state.IsPublishableVernacular || state.NoteTag != null)
+				if (!state.IsPublishable || state.NoteTag != null)
 					continue; // skip note text tokens and anything non-publishable
 				if (state.CharTag != null && _furtherInlineIgnorees.Contains(state.CharTag.Marker))
 					continue; // skip figure tokens
@@ -206,7 +207,7 @@ namespace HearThis.Script
 					{
 						// If we've been collecting chapter info and we're starting a new paragraph that we'll need to write out
 						// then we need to emit our chapter string first.
-						// [\cl and \cp have TextProperty paragraph, and IsPublishableVernacular is true,
+						// [\cl and \cp have TextProperty paragraph, and IsPublishable is true,
 						// but they DON'T have TextProperty Vernacular!]
 						if (collectingChapterInfo && state.ParaTag.TextProperties.HasFlag(TextProperties.scVernacular))
 						{
@@ -369,11 +370,8 @@ namespace HearThis.Script
 			// Enhance: GJM Eventually, hopefully, we can base this on a new 'for-audio'
 			// flag in TextProperties.
 			var isPublishable = tag.TextProperties.HasFlag(TextProperties.scPublishable);
-			var isVernacular = tag.TextProperties.HasFlag(TextProperties.scVernacular);
 			var isParagraph = tag.TextProperties.HasFlag(TextProperties.scParagraph);
-			if (isParagraph && isPublishable && isVernacular && !_furtherParagraphIgnorees.Contains(tag.Marker))
-				return true;
-			if (isParagraph && isPublishable && (tag.Marker == "cl" || tag.Marker == "cp"))
+			if (isParagraph && isPublishable && !_furtherParagraphIgnorees.Contains(tag.Marker))
 				return true;
 			return tag.TextProperties.HasFlag(TextProperties.scChapter);
 		}