+++ src/HearThisTests/ParserStateStub.cs
@@ -12,12 +12,12 @@ namespace HearThisTests
 		public HashSet<string> ParaMarkers = new HashSet<string>(new[] { "mt", "mt1", "mt2", "ip", "im", "ms", "imt", "q1", "s", "s1", "c", "cl", "cp", "p", "h", "r", "toc1" });
 		public HashSet<string> ParaMarkersNonReadable = new HashSet<string>(new[] { "rem" });
 
-		public HashSet<string> CharMarkers = new HashSet<string>(new[] { "bk", "fig", "rq" });
+		public HashSet<string> CharMarkers = new HashSet<string>(new[] { "bk", "fig", "rq", "tl" });
 
 		public void UpdateState(List<UsfmToken> tokenList, int tokenIndex)
 		{
 			ParaStart = false;
-			IsPublishableVernacular = false;
+			IsPublishable = false;
 
 			var marker = tokenList[tokenIndex].Marker;
 
@@ -57,13 +57,13 @@ namespace HearThisTests
 					case "cp":
 						AddTextProperty(ParaTag, TextProperties.scParagraph);
 						AddTextProperty(ParaTag, TextProperties.scPublishable);
-						IsPublishableVernacular = true;
+						IsPublishable = true;
 						break;
 					default:
 						AddTextProperty(ParaTag, TextProperties.scParagraph);
 						AddTextProperty(ParaTag, TextProperties.scPublishable);
-						AddTextProperty(ParaTag, TextProperties.scVernacular);
-						IsPublishableVernacular = true;
+						AddTextProperty(ParaTag, TextProperties.scVernacular); // This is not really required, but it corresponds to what would be in usfm.sty
+						IsPublishable = true;
 						ParaStart = true;
 						break;
 				}
@@ -84,7 +84,7 @@ namespace HearThisTests
 		public ScrTag CharTag { get; private set; }
 		public ScrTag ParaTag { get; internal set; }
 		public bool ParaStart { get; private set; }
-		public bool IsPublishableVernacular { get; private set; }
+		public bool IsPublishable { get; private set; }
 
 		private void AddTextProperty(ScrTag srcTag, TextProperties property)
 		{