+++ DistFiles/ReleaseNotes.md
@@ -51,6 +51,9 @@ the recorded files to that format, if necessary.
 
 # Release Notes
 
+## 3 January 2018
+- Fixed bug to add support for transliterated text (in \tl field). This change was incorporated into product for both Paratext 7 and Paratext 8.
+
 ## February 2017
 Fixed bug that cause HearThis to crash when not connected to the Internet.